Refactor kubernetes admission plugins
